requirements.txt
# --- API layer ---
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1
python-dotenv==1.0.1

# --- Vector DB + embeddings ---
qdrant-client==1.15.1
fastembed==0.7.3
numpy<2.3

# --- LLM (OpenRouter via OpenAI-compatible SDK) ---
openai==1.59.6

# --- Web search ---
tavily-python==0.5.1

# --- Scraping & chunking ---
requests==2.32.3
beautifulsoup4==4.12.3
markdownify==0.14.1
lxml==5.3.0

# --- Frontend ---
streamlit==1.41.1

# --- HTTP client used by frontend ---
httpx==0.28.1
